+++ services/services.go
@@ -73,6 +73,14 @@ func UnopaqueByte(compressed string, secretKey *[32]byte) ([]byte, error) {
 // Duration returns a friendly duration (with the insignificant bits rounded
 // off).
 func Duration(d time.Duration) string {
+	if d > 10*time.Second {
+		d2 := (d / (100 * time.Millisecond)) * (100 * time.Millisecond)
+		return d2.String()
+	}
+	if d > time.Second {
+		d2 := (d / (10 * time.Millisecond)) * (10 * time.Millisecond)
+		return d2.String()
+	}
 	d2 := (d / (100 * time.Microsecond)) * (100 * time.Microsecond)
 	return d2.String()
 }
@@ -84,4 +92,26 @@ func TruncateSid(sid string) string {
 		return sid[:8]
 	}
 	return sid
+}
+
+// FriendlyDate returns a friendlier version of the date.
+func FriendlyDate(t time.Time) string {
+	return friendlyDate(t, time.Now().UTC())
+}
+
+func friendlyDate(t time.Time, utcnow time.Time) string {
+	now := utcnow.In(t.Location())
+	y, m, d := now.Date()
+	if d == t.Day() && m == t.Month() && y == t.Year() {
+		return t.Format("3:04pm")
+	}
+	y1, m1, d1 := now.Add(-24 * time.Hour).Date()
+	if d1 == t.Day() && m1 == t.Month() && y1 == t.Year() {
+		return t.Format("Yesterday, 3:04pm")
+	}
+	// if the same year, return the day
+	if y == t.Year() {
+		return t.Format("3:04pm, January 2")
+	}
+	return t.Format("3:04pm, January 2, 2006")
 }